server/requirements.txt
# Multi-HRI Web Server Requirements
flask
flask-socketio
pyyaml
requests
pyzmq
eventlet
